# Possible Speedup: Move Normalization to GPU