Image normalization (cast to float and divide by 255) on GPU for a ~2x speedup when using larger history counts

Per sample timings on a mediocre machine (448x224 raster, 30 history frames):
77.9 ms ± 2.09 ms per loop (mean ± std. dev. of 7 runs, 10 loops each) with normalization (/255) on CPU
49.9 ms ± 2.61 ms per loop (mean ± std. dev. of 7 runs, 10 loops each) without normalization on CPU (must be done on GPU)