@lucabergamini 2021-01-14
Although the speedup is massive this is quite a breaking change for the codebase:
- old models won't work after this unless you patch all the lines where data is forwarded to them;
- it's very easy to forget to apply this normalisation as there is no general rule about where to do it (if it's not done at data level).