@pascal-pfeiffer 2021-01-21
I could think of a newly introduced config flag for gpu_normalization. 
- If true, the normalization would be done on the GPU and the user would benefit from the speedup. 
- If false (default value) nothing would change and the normalization would be done on the CPU